[DEBUG] Ensure pools are allocated before access.
EntityManager.componentId now ensures that the returned ID corresponds
to an allocated pool.

This fixes a bug where `unregister` would increment the pool count
without allocating a new pool, causing a successive `register` to
segfault.

This is an attempt to head off other similar bugs by avoiding the
possibility of accessing an un-allocated pool.