[DEBUG] Don't fail emitting event with no handler.
Check that there is an entry in the handlers table for the given event
id before trying to pass it on the the handlers.

The previous check (length == 0) was only sufficient to catch the case
where no handlers of _any_ type were registered. It would crash if at
least one kind of handler was registered, but there were no handlers
for the emitted event type.