[DEV] Automatic event subscription for systems.
Add an AutoSubscribe parameter to SystemManager.register/unregister.
When yes (the default), automatically subscribe the newly registered
system to any event E for which it implements Receiver!E.
Similarly, this applies to unregister and unsubscribing from events.

This will not work if the argument passed is an ISystem -- it must be
the full type implementing the Receiver interfaces.

This avoids the finicky business of adding new event receivers and
forgetting to subscribe them to the events, as now it is implicit in
registering the system.